Delete the actulas file
