+++ src/stories/containers/transparency-report/transparency-forecast/transparency-forecast-2.tsx
@@ -9,7 +9,7 @@ import { AdvancedInnerTable } from '../../../components/advanced-inner-table/adv
 import { CustomLink } from '../../../components/custom-link/custom-link';
 import { Tabs } from '../../../components/tabs/tabs';
 import { TransparencyEmptyTable } from '../placeholders/transparency-empty-table';
-import { LinkDescription } from '../transparency-actuals/transparency-actuals-2';
+import { LinkDescription } from '../transparency-actuals/transparency-actuals';
 import { Title } from '../transparency-report';
 import { useTransparencyForecastMvvm2 } from './transparency-forecast-2.mvvm';
 import type { BudgetStatementDto } from '../../../../core/models/dto/core-unit.dto';
